toolkit/components/extensions/test/mochitest/test_chrome_ext_downloads_saveAs.html

<!doctype html>
<html>
<head>
  <title>Test downloads.download() saveAs option</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SpawnTask.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script src="chrome_head.js"></script>
  <script src="head.js"></script>
  <link rel="stylesheet" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="text/javascript">
"use strict";

Cu.import("resource://gre/modules/FileUtils.jsm");

let directory;

add_task(async function setup() {
  directory = FileUtils.getDir("TmpD", ["downloads"]);
  directory.createUnique(Ci.nsIFile.DIRECTORY_TYPE, FileUtils.PERMS_DIRECTORY);
  info(`Using download directory ${directory.path}`);

  await SpecialPowers.pushPrefEnv({"set": [
    ["browser.download.folderList", 2],
    ["browser.download.dir", directory.path],
  ]});

  SimpleTest.registerCleanupFunction(async () => {
    await SpecialPowers.popPrefEnv();
    directory.remove(true);
  });
});

add_task(async function test_downloads_saveAs() {
  const file = directory.clone();
  file.append("file_download.txt");

  const unique = directory.clone();
  unique.append("file_download(1).txt");

  const {MockFilePicker} = SpecialPowers;
  MockFilePicker.init(window);

  MockFilePicker.showCallback = fp => {
    let file = directory.clone();
    file.append(fp.defaultString);
    MockFilePicker.setFiles([file]);
  };

  function background() {
    const url = URL.createObjectURL(new Blob(["file content"]));
    browser.test.onMessage.addListener(async filename => {
      try {
        let id = await browser.downloads.download({
          url,
          filename,
          saveAs: true,
          conflictAction: "uniquify",
        });
        browser.downloads.onChanged.addListener(delta => {
          if (delta.id == id && delta.state.current === "complete") {
            browser.test.sendMessage("done", {ok: true, id});
          }
        });
      } catch ({message}) {
        browser.test.sendMessage("done", {ok: false, message});
      }
    });
    browser.test.sendMessage("ready");
  }

  const manifest = {background, manifest: {permissions: ["downloads"]}};
  const extension = ExtensionTestUtils.loadExtension(manifest);

  await extension.startup();
  await extension.awaitMessage("ready");

  // Test basic saveAs functionality.
  MockFilePicker.returnValue = MockFilePicker.returnOK;

  extension.sendMessage("file_download.txt");
  let result = await extension.awaitMessage("done");
  ok(result.ok, "downloads.download() works with saveAs");

  ok(file.exists(), "the file exists.");
  is(file.fileSize, 12, "downloaded file is the correct size");

  // Test exisisting file with uniquify.
  MockFilePicker.returnValue = MockFilePicker.returnOK;

  extension.sendMessage("file_download.txt");
  result = await extension.awaitMessage("done");
  ok(result.ok, "downloads.download() works with saveAs and uniquify");

  ok(unique.exists(), "the file exists.");
  is(unique.fileSize, 12, "downloaded file is the correct size");
  unique.remove(false);

  // Test canceled saveAs for an existing file and uniquify.
  MockFilePicker.returnValue = MockFilePicker.returnCancel;

  extension.sendMessage("file_download.txt");
  result = await extension.awaitMessage("done");

  ok(!result.ok, "download rejected if the user cancels the dialog");
  is(result.message, "Download canceled by the user", "with the correct message");

  ok(!unique.exists(), "unique file not left after SaveAs canceled.");
  file.remove(false);

  // Test the user canceling the save dialog.
  MockFilePicker.returnValue = MockFilePicker.returnCancel;

  extension.sendMessage("file_download.txt");
  result = await extension.awaitMessage("done");

  ok(!result.ok, "download rejected if the user cancels the dialog");
  is(result.message, "Download canceled by the user", "with the correct message");
  ok(!file.exists(), "file was not downloaded");

  await extension.unload();
  MockFilePicker.cleanup();
});

</script>

</body>
</html>
